backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai-agents[litellm]>=0.6.2
python-dotenv==1.0.1
asyncpg>=0.29.0
better-auth==0.0.1b11
pyjwt[crypto]==2.10.1
python-multipart==0.0.20
